+++ src/test/resources/html/iFrameWithEverything.html
@@ -1,2 +1,2 @@
-<iframe id="iFrame" class="red blue green" allow="geolocation" allowfullscreen height="200" loading="eager" name="thing" referrerpolicy="origin" sandbox="allow-forms" src="https://www.google.com" srcdoc="something" width="200">
+<iframe id="iFrame" class="red blue green" height="200" src="https://www.google.com" width="200" allow="geolocation" allowfullscreen loading="eager" name="thing" referrerpolicy="origin" sandbox="allow-forms" srcdoc="something">
 </iframe>
