+++ src/test/resources/html/imgWithEverything.html
@@ -1,1 +1,1 @@
-<img id="aGoodBoy" class="cute pet photo" alt="A cool doggo" crossorigin="anonymous" height="200" ismap loading="eager" longdesc="/doggo" referrerpolicy="origin" sizes="something" src="./doggo.png" srcset="something-else" usemap="#some-map" width="200">
+<img id="aGoodBoy" class="cute pet photo" height="200" src="./doggo.png" width="200" alt="A cool doggo" crossorigin="anonymous" ismap loading="eager" longdesc="/doggo" referrerpolicy="origin" sizes="something" srcset="something-else" usemap="#some-map">
