+++ src/test/resources/html/simpleImg.html
@@ -1,1 +1,1 @@
-<img alt="A cool doggo" src="./doggo.png">
+<img src="./doggo.png" alt="A cool doggo">
